+++ src/components/Toolbars/Toolbars.js
@@ -1,98 +1,7 @@
-import { useState, useRef, PureComponent, createRef } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleDown } from "@fortawesome/free-solid-svg-icons";
-import { getDataFromPDF, getDataFromJSON } from "./utils.js";
-
-//PDF import
-function ImportTools({ onTextUpdate, onWidthUpdate, onJSONImport }) {
-  const fileInputRef = useRef();
-  const [type, setType] = useState("");
-  const [hovering, setHovering] = useState(false);
-
-  const importFile = (e) => {
-    if (!fileInputRef.current.value) {
-      console.log("no file picked");
-      return;
-    } else {
-      const file = fileInputRef.current.files[0];
-      if (type === "pdf") {
-        getDataFromPDF(file).then(({ width, bullets }) => {
-          onWidthUpdate(width);
-          onTextUpdate(bullets);
-        });
-      } else if (type === "json") {
-        getDataFromJSON(file).then((data) => onJSONImport(data));
-      }
-
-      fileInputRef.current.value = "";
-    }
-  };
-
-  const menuState = hovering ? "is-active" : "";
-
-  return (
-    <div className={"dropdown " + menuState}>
-      <input
-        type="file"
-        onChange={importFile}
-        style={{ display: "none" }}
-        ref={fileInputRef}
-      ></input>
-      <div className="dropdown-trigger">
-        <div className="buttons has-addons">
-          <button
-            className="button"
-            onClick={() => {
-              setType("pdf");
-              fileInputRef.current.click();
-            }}
-          >
-            Import
-          </button>
-          <button
-            className="button"
-            onClick={() => setHovering(!hovering)}
-            aria-haspopup="true"
-            aria-controls="import-menu"
-          >
-            <span className="icon">
-              <FontAwesomeIcon icon={faAngleDown} />
-            </span>
-          </button>
-        </div>
-      </div>
-      <div
-        className="dropdown-menu"
-        id="import-menu"
-        role="menu"
-        onMouseLeave={() => setHovering(false)}
-      >
-        <div className="dropdown-content">
-          <a
-            href="?#"
-            className="dropdown-item"
-            onClick={() => {
-              setType("pdf");
-              fileInputRef.current.click();
-            }}
-          >
-            PDF
-          </a>
-          <a
-            href="?#"
-            className="dropdown-item"
-            onClick={() => {
-              setType("json");
-              fileInputRef.current.click();
-            }}
-          >
-            JSON
-          </a>
-        </div>
-      </div>
-    </div>
-  );
-}
+import ImportTools from "./ImportTools";
+import SaveTools from "./SaveTools";
 // form width, space optimization, select text
 function OutputTools({ onOptimChange, enableOptim, onWidthUpdate, width }) {
   const widthAWD = 202.321;
@@ -163,88 +72,7 @@ function InputTools({ onTextNorm }) {
     </button>
   );
 }
-// saving settings
-function SaveTools({ getSavedSettings }) {
-  const [hovering, setHovering] = useState(false);
-  const exportRef = useRef();
-  const menuState = hovering ? "is-active" : "";
-  const saveSettings = () => {
-    const settings = getSavedSettings();
-    //JSON stringifying an array for future growth
-    const storedData = JSON.stringify([settings]);
-    try {
-      localStorage.setItem("bullet-settings", storedData);
-      console.log(
-        "saved settings/data to local storage with character length " +
-          storedData.length
-      );
-    } catch (err) {
-      if (err.name === "SecurityError") {
-        alert(
-          "Sorry, saving to cookies does not work using the file:// interface and/or your browser's privacy settings"
-        );
-      } else {
-        throw err;
-      }
-    }
-  };
-  const exportSettings = () => {
-    const settings = getSavedSettings();
-    //JSON stringifying an array for future growth
-    const storedData = JSON.stringify([settings]);
-    const dataURI =
-      "data:application/JSON;charset=utf-8," + encodeURIComponent(storedData);
-    exportRef.current.href = dataURI;
-    exportRef.current.click();
-    console.log(
-      "exported settings/data to JSON file with character length " +
-        storedData.length
-    );
-  };
 
-  return (
-    <div className={"dropdown " + menuState}>
-      <div className="dropdown-trigger">
-        <div className="buttons has-addons">
-          <button className="button" onClick={saveSettings}>
-            Save{" "}
-          </button>
-          <button
-            className="button"
-            aria-haspopup="true"
-            aria-controls="save-menu"
-          >
-            <span className="icon" onClick={() => setHovering(!hovering)}>
-              <FontAwesomeIcon icon={faAngleDown} />
-            </span>
-          </button>
-        </div>
-      </div>
-      <div
-        className="dropdown-menu"
-        id="save-menu"
-        role="menu"
-        onMouseLeave={() => setHovering(false)}
-      >
-        <div className="dropdown-content">
-          <a href="?#" className="dropdown-item" onClick={saveSettings}>
-            Cookie
-          </a>
-          <a href="?#" className="dropdown-item" onClick={exportSettings}>
-            JSON
-          </a>
-        </div>
-      </div>
-
-      <a
-        href="?#"
-        style={{ display: "none" }}
-        download="settings.json"
-        ref={exportRef}
-      ></a>
-    </div>
-  );
-}
 function Logo() {
   return (
     <h1 className="title">
